kubeConfigId/clusterId 변수형을 Long형으로 통일

+++ src/main/java/kr/co/strato/adapter/k8s/common/model/YamlApplyParam.java
@@ -1,6 +1,10 @@
 package kr.co.strato.adapter.k8s.common.model;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -8,7 +12,7 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class YamlApplyParam {
-    private Integer kubeConfigId;
+    private Long kubeConfigId;
     private String yaml;
 
 }
