+++ src/main/java/kr/co/strato/adapter/k8s/common/proxy/NonNamespaceProxy.java
@@ -1,9 +1,16 @@
 package kr.co.strato.adapter.k8s.common.proxy;
 
+import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+
 import kr.co.strato.adapter.k8s.common.model.ClusterResourceInfo;
 import kr.co.strato.adapter.k8s.common.model.ResourceListSearchInfo;
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.web.bind.annotation.*;
 
 @FeignClient(value="NonNamespace", url = "${service.kubernetes-interface.url}")
 public interface NonNamespaceProxy {
@@ -17,13 +24,13 @@ public interface NonNamespaceProxy {
     @GetMapping("/nonNamespace/{resourceType}/detailStr")
     public @ResponseBody String getResource(
             @PathVariable("resourceType") String resourceType,
-            @RequestParam("kubeConfigId") Integer kubeConfigId,
+            @RequestParam("kubeConfigId") Long kubeConfigId,
             @RequestParam("name") String name);
 
     @GetMapping("/nonNamespace/{resourceType}/yaml")
     public String getResourceYaml(
             @PathVariable("resourceType") String resourceType,
-            @RequestParam("kubeConfigId") Integer kubeConfigId,
+            @RequestParam("kubeConfigId") Long kubeConfigId,
             @RequestParam("name") String name);
 
     @DeleteMapping("/nonNamespace/{resourceType}")
